# docs/openapi.yaml
basePath: /proxy/v1
definitions:
  common.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  common.PaginatedSearchResponse:
    properties:
      metadata:
        $ref: '#/definitions/common.PaginationMetadata'
      results:
        items:
          $ref: '#/definitions/models.SearchItem'
        type: array
    type: object
  common.PaginationMetadata:
    properties:
      page:
        type: integer
      total_pages:
        type: integer
      total_results:
        type: integer
    type: object
...
  models.AlbumResponse:
    properties:
      album_type:
        type: string
      authors:
        items:
          $ref: '#/definitions/models.Author'
        type: array
      duration_minutes:
        type: integer
      external_url:
        type: string
      id:
        type: string
      image_url:
        type: string
      images:
        items:
          $ref: '#/definitions/models.ImageEntry'
        type: array
      release_date:
        type: string
      title:
        type: string
      total_tracks:
        type: integer
      tracks:
        items:
          $ref: '#/definitions/models.Track'
        type: array
      type:
        type: string
    type: object
  models.Author:
    properties:
      name:
        type: string
      type:
        type: string
    type: object
  models.BookResponse:
    properties:
      authors:
        items:
          $ref: '#/definitions/models.Author'
        type: array
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      images:
        items:
          $ref: '#/definitions/models.ImageEntry'
        type: array
      pages:
        type: integer
      release_date:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
  models.ImageEntry:
    properties:
      image_url:
        type: string
      size:
        $ref: '#/definitions/models.ImageSize'
      type:
        $ref: '#/definitions/models.ImageType'
    type: object
  models.ImageSize:
    enum:
    - standard
    - original
    type: string
    x-enum-varnames:
    - ImageSizeStandard
    - ImageSizeOriginal
  models.ImageType:
    enum:
    - poster
    - gallery
    type: string
    x-enum-varnames:
    - ImageTypePoster
    - ImageTypeGallery
...
  models.SearchItem:
    properties:
      authors:
        items:
          $ref: '#/definitions/models.Author'
        type: array
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      original_title:
        type: string
      release_date:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
  models.Track:
    properties:
      authors:
        items:
          $ref: '#/definitions/models.Author'
        type: array
      duration_seconds:
        type: integer
      external_url:
        type: string
      id:
        type: string
      title:
        type: string
      track_number:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    REST API proxy that aggregates metadata from TMDB, IGDB, Spotify, and OpenLibrary.
    Authentication: send API key via X-Api-Key header or Authorization: Bearer <token>.
    Rate limit: 60 requests/minute on protected endpoints.
  title: Denn Proxy API
  version: "1.0"
paths:
  /albums/{id}:
    get:
      parameters:
      - description: Spotify album ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AlbumResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Get album details
      tags:
      - Albums
  /albums/bulk:
    get:
      description: Fetches full details for up to 20 albums in parallel. Album IDs
        are Spotify string IDs. Failed IDs are silently omitted.
      parameters:
      - description: Comma-separated Spotify album IDs (max 20)
        example: 5K79FLRUCSysQnVESLcTdb,3RQQmkQEvNCY4prGKE6oc5
        in: query
        name: ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AlbumResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Get multiple albums by ID
      tags:
      - Albums
  /albums/search:
    get:
      description: Singles are excluded from results (only albums and EPs are returned).
      parameters:
      - description: Search term
        in: query
        name: query
        required: true
        type: string
      - default: 1
        description: Page number (min 1)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Results per page (max 50)
        in: query
        maximum: 50
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.PaginatedSearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Search albums
      tags:
      - Albums
  /albums/trending:
    get:
      description: Sourced from Spotify Charts.
      parameters:
      - default: 1
        description: Page number (min 1)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Results per page (max 50)
        in: query
        maximum: 50
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.PaginatedSearchResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Get trending albums
      tags:
      - Albums
  /books/{id}:
    get:
      parameters:
      - description: OpenLibrary work ID (e.g. OL27448W)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Get book details
      tags:
      - Books
  /books/bulk:
    get:
      description: Fetches full details for up to 20 books in parallel. Book IDs are
        OpenLibrary work IDs (e.g. OL27448W). Failed IDs are silently omitted.
      parameters:
      - description: Comma-separated OpenLibrary work IDs (max 20)
        example: OL27448W,OL82563W
        in: query
        name: ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BookResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Get multiple books by ID
      tags:
      - Books
  /books/search:
    get:
      parameters:
      - description: Search term
        in: query
        name: query
        required: true
        type: string
      - default: 1
        description: Page number (min 1)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Results per page (max 50)
        in: query
        maximum: 50
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.PaginatedSearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Search books
      tags:
      - Books
  /books/trending:
    get:
      description: Sourced from OpenLibrary bestsellers with client-side pagination.
      parameters:
      - default: 1
        description: Page number (min 1)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Results per page (max 50)
        in: query
        maximum: 50
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.PaginatedSearchResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - BearerAuth: []
      summary: Get trending books
      tags:
      - Books
...
securityDefinitions:
  ApiKeyHeader:
    description: API key passed in the X-Api-Key header
    in: header
    name: X-Api-Key
    type: apiKey
  BearerAuth:
    description: 'API key passed as a Bearer token: "Authorization: Bearer <key>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"